added cheks

+++ backend/regular_payment/regular_payment_service.cpp
@@ -22,28 +22,80 @@ vector<RegularPaymentDto> RegularPaymentService::get_all_by_user(const TokenDto 
 
 void RegularPaymentService::add(const RegularPaymentCreateDto &regular_payment_create_dto) const {
     const Account* account = _get_account(regular_payment_create_dto._token);
+    time_t current_time = time(nullptr);
+
+    if(regular_payment_create_dto._next_time <= current_time){
+        throw Exception("Next payment should be in the future");
+    }
+
+    if(regular_payment_create_dto._period_sec <= 0){
+        throw Exception("Period should be positive");
+    }
+
+    Optional<Account> optional_account = _account_repository.get_by_card_number(regular_payment_create_dto._target_card);
+    if(optional_account.is_empty()){
+        throw Exception("Target account does not exists");
+    }
+
+    if(regular_payment_create_dto._sum <= 0){
+        throw Exception("Sum should be positive");
+    }
+
     _regular_payment_repository.add(RegularPayment{0,
                                                    account->_card_number,
                                                    regular_payment_create_dto._period_sec,
                                                    regular_payment_create_dto._target_card,
-                                                   regular_payment_create_dto._next_time - regular_payment_create_dto._period_sec,
+                                                   regular_payment_create_dto._next_time,
                                                    regular_payment_create_dto._sum});
 
 }
 
 void RegularPaymentService::update(const RegularPaymentUpdateDto &regular_payment_update_dto) const {
     const Account* account = _get_account(regular_payment_update_dto._token);
+    time_t current_time = time(nullptr);
+
+    Optional<RegularPayment> optional_payment = _regular_payment_repository.get_by_id(regular_payment_update_dto._payment_id);
+    if(optional_payment.is_empty()){
+        throw Exception("Attempt to change non-existent regular payment");
+    }
+
+    if(regular_payment_update_dto._next_time <= current_time){
+        throw Exception("Next payment should be in the future");
+    }
+
+    if(regular_payment_update_dto._period_sec <= 0){
+        throw Exception("Period should be positive");
+    }
+
+    Optional<Account> optional_account = _account_repository.get_by_card_number(regular_payment_update_dto._target_card);
+    if(optional_account.is_empty()){
+        throw Exception("Target account does not exists");
+    }
+
+    if(regular_payment_update_dto._sum <= 0){
+        throw Exception("Sum should be positive");
+    }
+
     _regular_payment_repository.update(RegularPayment{regular_payment_update_dto._payment_id,
                                                       account->_card_number,
                                                       regular_payment_update_dto._period_sec,
                                                       regular_payment_update_dto._target_card,
-                                                      regular_payment_update_dto._next_time - regular_payment_update_dto._period_sec,
+                                                      regular_payment_update_dto._next_time,
                                                       regular_payment_update_dto._sum
                                                       });
 }
 
 void RegularPaymentService::remove(const RegularPaymentDeleteDto &regular_payment_delete_dto) const {
-    _get_account(regular_payment_delete_dto._token);
+    const Account* account = _get_account(regular_payment_delete_dto._token);
+    Optional<RegularPayment> optional_payment = _regular_payment_repository.get_by_id(regular_payment_delete_dto._payment_id);
+    if(optional_payment.is_empty()){
+        throw Exception("Regular payment does not exists");
+    }
+
+    if(account->_card_number != optional_payment.get()->_account_card){
+        throw Exception("Attempt to delete someone else's regular payment");
+    }
+
     _regular_payment_repository.remove(regular_payment_delete_dto._payment_id);
 }
 
@@ -51,7 +103,7 @@ void RegularPaymentService::remove(const RegularPaymentDeleteDto &regular_paymen
 vector<RegularPayment> RegularPaymentService::get_to_be_paid() const {
     time_t current_time = time(nullptr);
     return _regular_payment_repository.get_list(Specification<RegularPayment>(
-            [&](const RegularPayment& rp){return rp._last_paid + rp._period_sec <= current_time;}
+            [&](const RegularPayment& rp){return rp._next_time + rp._period_sec <= current_time;}
             ));
 }
